+++ changelog.md
@@ -4,6 +4,7 @@
 * Quartz arrows now deal bonus piercing damage instead of flying faster and farther
 * Fixed part breaker dropping multiple items that already dropped naturally
   * E.g. Wither Skeletons can no longer drop 2 skulls
+* Serene Seasons' seasonal_crops is no longer set to false, as it's done in ITR
 
 ## 1.12.2
 * Updated advancements